Organise test project
Sample view models are now in the samples folder, setup & verification tests are in the root project directory
